refactor: rename file BasicInformation.tsx to PersonalDetails.tsx
